Changed return value of CpuData::() from u16 to PSW&.

+++ gtest/kdf11_a/kdf11_a_pswregistertest.cpp
@@ -6,7 +6,7 @@ TEST (KD_PSWREGISTERTEST, PswCanBeRead)
 {
     KD_PSW psw {0};
 
-    ASSERT_EQ (psw, 0);
+    ASSERT_EQ ((u16) psw, 0);
 }
 
 TEST (KD_PSWREGISTERTEST, ConditionCodesCanBeCleared)
@@ -14,7 +14,7 @@ TEST (KD_PSWREGISTERTEST, ConditionCodesCanBeCleared)
     KD_PSW psw {0177777};
     psw.clearConditionCodes (017);
 
-    ASSERT_EQ (psw, 0177760);
+    ASSERT_EQ ((u16) psw, 0177760);
 }
 
 TEST (KD_PSWREGISTERTEST, ConditionCodesCanBeSet)
@@ -22,7 +22,7 @@ TEST (KD_PSWREGISTERTEST, ConditionCodesCanBeSet)
     KD_PSW psw {0};
     psw.setConditionCodes (017);
 
-    ASSERT_EQ (psw, 017);
+    ASSERT_EQ ((u16) psw, 017);
 }
 
 TEST (KD_PSWREGISTERTEST, TraceBitSet)
@@ -36,7 +36,7 @@ TEST (KD_PSWREGISTERTEST, PriorityLevel)
     KD_PSW psw {0};
     psw.setPriorityLevel (7);
 
-    ASSERT_EQ (psw, 0000340);
+    ASSERT_EQ ((u16) psw, 0000340);
     ASSERT_EQ (psw.priorityLevel (), 7);
 }
